refactor WizgovWidget: remove WizgovWidgetClient, integrate script loading directly, and simplify props handling
