+++ packages/components/src/interfaces/internal/index.ts
@@ -42,8 +42,4 @@ export type {
   GoogleTagManagerBodyProps,
 } from "./GoogleTagManager"
 export type { AttrsDirProps } from "./AttrsDir"
-export type {
-  WizgovWidgetProps,
-  WizgovWidgetClientProps,
-  WizgovProps,
-} from "./Wizgov"
+export type { WizgovWidgetProps, WizgovProps } from "./Wizgov"
